Reogranisation of the API structure
